PP-11393 Remove duplicate payment intent model
StripePaymentIntentResponse and StripePaymentIntent both represent a
Payment Intent in an API response from Stripe. Combine these so that
we only have one model.
