+++ src/main/java/uk/gov/pay/connector/gateway/stripe/StripeAuthorisationResponse.java
@@ -4,8 +4,8 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import uk.gov.pay.connector.gateway.model.Gateway3dsRequiredParams;
 import uk.gov.pay.connector.gateway.model.response.BaseAuthoriseResponse;
+import uk.gov.pay.connector.gateway.stripe.json.StripePaymentIntent;
 import uk.gov.pay.connector.gateway.stripe.response.Stripe3dsRequiredParams;
-import uk.gov.pay.connector.gateway.stripe.response.StripePaymentIntentResponse;
 import uk.gov.pay.connector.gateway.stripe.response.StripePaymentMethodResponse;
 import uk.gov.service.payments.commons.model.CardExpiryDate;
 
@@ -52,12 +52,12 @@ public class StripeAuthorisationResponse implements BaseAuthoriseResponse {
         this.cardExpiryDate = cardExpiryDate;
     }
 
-    public static StripeAuthorisationResponse of(StripePaymentIntentResponse stripePaymentIntent) {
+    public static StripeAuthorisationResponse of(StripePaymentIntent stripePaymentIntent) {
         return new StripeAuthorisationResponse(
                 stripePaymentIntent.getId(),
                 stripePaymentIntent.getAuthoriseStatus().orElse(null),
                 stripePaymentIntent.getRedirectUrl().orElse(null),
-                stripePaymentIntent.getStringifiedOutcome(),
+                stripePaymentIntent.stringify(),
                 stripePaymentIntent.getCustomerId(),
                 stripePaymentIntent.getPaymentMethod().getId(),
                 extractCardExpiryDate(stripePaymentIntent).orElse(null)
@@ -108,7 +108,7 @@ public class StripeAuthorisationResponse implements BaseAuthoriseResponse {
         return Optional.of(cardExpiryDate);
     }
 
-    private static Optional<CardExpiryDate> extractCardExpiryDate(StripePaymentIntentResponse stripePaymentIntent) {
+    private static Optional<CardExpiryDate> extractCardExpiryDate(StripePaymentIntent stripePaymentIntent) {
         return stripePaymentIntent.getPaymentMethod().getExpanded()
                 .flatMap(StripePaymentMethodResponse::getCard)
                 .map(card -> {
