+++ src/test/resources/templates/stripe/notification_payment_intent_payment_failed.json
@@ -20,6 +20,7 @@
       "canceled_at": null,
       "cancellation_reason": null,
       "capture_method": "manual",
+      "payment_method": "pm_1FHESdEZsufgnuO0bzghQoZ2",
       "charges": {
         "object": "list",
         "data": [
